StateManager: remove BaseStateManager
