+++ packages/statemanager/src/index.ts
@@ -1,4 +1,3 @@
-export { BaseStateManager } from './baseStateManager'
 export { Cache, CacheOpts, deleteCb, getCb, putCb } from './cache'
 export { EthersStateManager, EthersStateManagerOpts } from './ethersStateManager'
 export { AccountFields, StateAccess, StateManager } from './interface'
